重构播放器加载分P和合集的逻辑；重做播放器内视频列表界面

+++ bili-api/src/main/kotlin/dev/aaa1115910/biliapi/http/BiliHttpApi.kt
@@ -162,7 +162,7 @@ object BiliHttpApi {
      * 获取视频详细信息
      */
     suspend fun getVideoInfo(
-        av: Int? = null,
+        av: Long? = null,
         bv: String? = null,
         sessData: String? = null
     ): BiliResponse<VideoInfo> = client.get("/x/web-interface/view") {
